fix: settings component referred in ReactNode
