+++ client/src/apis/todo.ts
@@ -26,7 +26,6 @@ const createTodo = async (todo: Todo): Promise<Todo> => {
 const updateTodo = async (todo: Todo): Promise<Todo> => {
     const req: TodoRequestModel = renameKeys(todo, { id: 'uuid' });
     const { data } = await http.post<Todo>('/todo/update', req);
-    console.log('data', data);
     return data;
 };
