+++ client/src/components/Settings/Settings.tsx
@@ -1,27 +1,10 @@
-import { Button, SettingsAbout, SettingsNav, SettingsProfile } from '@/components';
+import { Button, SettingsAbout, SettingsNav, SettingsNavTabs, SettingsProfile } from '@/components';
+import { SettingNavContent, SettingsNavs } from '@/components/Settings/SettingsNavTabs';
 import { IconClose } from '@douyinfe/semi-icons';
 import { ActionIcon, Modal, Button as MButton } from '@mantine/core';
 import { IconMessageCircle, IconPacman, IconPhoto, IconSettings, IconUser } from '@tabler/icons';
 import clsx from 'clsx';
 import { ReactNode, useState } from 'react';
-
-export const enum SettingsNavTabs {
-    Profile = 'Profile',
-    About = 'About'
-}
-
-export const SettingsNavs: SettingsNav[] = [
-    {
-        label: SettingsNavTabs.Profile,
-        icon: <IconUser />,
-        component: <SettingsProfile />
-    },
-    {
-        label: SettingsNavTabs.About,
-        icon: <IconPacman />,
-        component: <SettingsAbout />
-    }
-];
 
 export const Settings = (props: { opened: boolean; onClose: () => void }) => {
     const [activeTab, setActiveTab] = useState<SettingsNavTabs>(SettingsNavTabs.Profile);
@@ -81,7 +64,7 @@ export const Settings = (props: { opened: boolean; onClose: () => void }) => {
                         )}
                     >
                         <div className={'tw-flex'}>
-                            {SettingsNavs.find((nav) => nav.label === activeTab)?.component}
+                            <SettingNavContent activeTab={activeTab} />
                         </div>
                     </div>
                 </div>
