+++ client/src/components/Settings/SettingsAbout.tsx
@@ -2,5 +2,5 @@ import { useTauriExtension } from '@/hooks';
 
 export const SettingsAbout = () => {
     const { openLink } = useTauriExtension();
-    return <></>;
+    return <>ohe</>;
 };
