+++ client/src/components/Settings/SettingsProfile.tsx
@@ -1,3 +1,12 @@
+import { Title, Card, Avatar } from '@mantine/core';
+
 export const SettingsProfile = () => {
-    return <>aaa</>;
+    return (
+        <div className={'flex'}>
+            <Title order={2}>Profile</Title>
+            <Card radius={'lg'} withBorder>
+                <Avatar radius={'xl'} size={'xl'} />
+            </Card>
+        </div>
+    );
 };
