+++ client/src/components/index.ts
@@ -15,10 +15,11 @@ export * from './Heros/CopyrightBar';
 export * from './Heros/SocialIcons';
 export * from './Heros/WelcomeTitle';
 export * from './Resizer/Resizer';
-export * from './Settings/SettingsAbout';
-export * from './Settings/SettingsProfile';
 export * from './Settings/SettingNav';
 export * from './Settings/Settings';
+export * from './Settings/SettingsAbout';
+export * from './Settings/SettingsNavTabs';
+export * from './Settings/SettingsProfile';
 export * from './Spacer/Spacer';
 export * from './Todo/TodoApp';
 export * from './Todo/TodoInput';
